% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing-functions.R
\name{bw.rot}
\alias{bw.rot}
\title{Silverman's rule-of-thumb bandwidth}
\usage{
bw.rot(x)
}
\arguments{
\item{x}{A numeric vector without non-finite values.}
}
\value{
A bandwidth that might be optimal for non-parametric density estimation of \code{x}.
}
\description{
A fail-safe function that would return a nice Silverman-like bandwidth suggestion for data for which the standard deviation might be NA or 0.
}
\examples{
set.seed(1)
bw.rot(rnorm(100)) # Should be 0.3787568 in R version 4.0.4
}
