% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing-functions.R
\name{DCV}
\alias{DCV}
\alias{LSCV}
\title{Density and Least-Squares cross-validation criteria}
\usage{
DCV(
  x,
  bw,
  kernel = "gaussian",
  same = FALSE,
  rescale = TRUE,
  parallel = FALSE,
  ncores = 2
)

LSCV(
  x,
  y,
  bw,
  kernel = "gaussian",
  same = FALSE,
  rescale = TRUE,
  parallel = FALSE,
  ncores = 2
)
}
\arguments{
\item{x}{A numeric vector or numeric matrix.}

\item{bw}{A numeric scalar, vector, or matrix of candidate bandwidths.}

\item{kernel}{Which kernel to use? Passed to \code{kernelWeights}.}

\item{same}{Logical: use the same bandwidth in all dimensions for multi-variate x?}

\item{rescale}{Passed to \code{kernelWeights}.}

\item{parallel}{Logical: use \code{parallel::mclapply} to speed up computation?}

\item{ncores}{The number of cores to use for parallel computation. Reset to 1 on Windows machines.}

\item{y}{A numeric vector of responses (dependent variable).}
}
\value{
For one-dimensional \code{x}, returns a numeric vector of the same length as \code{length(bw)}.
For matrix \code{x}, returns a numeric vector of the same length as \code{nrow(bw).}
}
\description{
Density cross-validation is performed based purely on the values of \code{x} for the kernel density estimator.
Least-squares cross-validation is performed for the Nadaraya-Watson estimator of E(Y|X).
}
\examples{
set.seed(1)
x <- rnorm(100)
y <- x^2 + rnorm(100)
plot(x, y, bty = "n")
bw.grid <- exp(seq(log(bw.rot(x) / 4), log(bw.rot(x) * 2), length.out = 101))
DCV.values <- DCV(x, bw = bw.grid)
DCV.values <- (DCV.values - min(DCV.values)) / diff(range(DCV.values))
LSCV.values <- LSCV(x = x, y = y, bw = bw.grid)
LSCV.values <- (LSCV.values - min(LSCV.values)) / diff(range(LSCV.values))
plot(bw.grid, DCV.values, bty = "n", xlab = "Bandwidth", type = "l")
lines(bw.grid, LSCV.values, col = "red")
}
