% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing-functions.R
\name{kernelMixedSmooth}
\alias{kernelMixedSmooth}
\title{Kernel smoothing with conditioning on discrete and continuous variables}
\usage{
kernelMixedSmooth(x, y, by, xgrid = NULL, bygrid = NULL, bw, ...)
}
\arguments{
\item{x}{A numeric vector or a numeric matrix of continuous predictors}

\item{y}{A numeric vector of responses (dependent variable).}

\item{by}{An integer defining the grouping (all possible unique combinations of discrete predictor values)}

\item{xgrid}{A numeric vector or numeric matrix with \code{ncol(xgrid) = ncol(x)} of points at which the density is estimated.}

\item{bygrid}{An integer defining the grouping for the grid (all possible unique combinations of discrete predictor values)}

\item{bw}{Bandwidth: a scalar or a vector of the same length as \code{ncol(x)}.}

\item{...}{Passed to kernelSmooth.}
}
\value{
A numeric vector of kernel density estimator of \code{x} evaluated at \code{xgrid}.
}
\description{
Kernel smoothing with conditioning on discrete and continuous variables
}
\seealso{
\code{\link{kernelSmooth}} for the underlying function.
}
