% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing-functions.R
\name{kernelWeights}
\alias{kernelWeights}
\title{Kernel weight matrix computation}
\usage{
kernelWeights(
  x,
  xgrid = NULL,
  bw,
  kernel = "gaussian",
  order = 2,
  rescale = TRUE,
  convolution = FALSE
)
}
\arguments{
\item{x}{A numeric vector or numeric matrix of predictors.}

\item{xgrid}{A numeric vector or numeric matrix with \code{ncol(xgrid) = ncol(x)}.}

\item{bw}{Bandwidth: a scalar or a vector of the same length as ncol(x).}

\item{kernel}{Passed to \code{kernelFun}.}

\item{order}{Passed to \code{kernelFun}.}

\item{rescale}{Passed to \code{kernelFun}.}

\item{convolution}{Passed to \code{kernelFun}.}
}
\value{
A numeric matrix of size \code{nrow(x) x nrow(xgrid)} with kernel weights.
}
\description{
Kernel weight matrix computation
}
\examples{
set.seed(1)
x <- sort(rnorm(30))
kw <- kernelWeights(x, xgrid = seq(-3, 3, 0.05), bw = bw.rot(x))
image(kw)
mycols <- rainbow(length(x), end = 0.7, v = 0.8)
plot(NULL, NULL, xlim = c(-3, 3), ylim = c(0, max(kw)), bty = "n", xlab = "x", ylab = "Weights")
for (i in 1:length(x)) lines(seq(-3, 3, 0.05), kw[i, ], col = mycols[i])
abline(v = x, col = mycols, lty = 3)
}
\seealso{
\code{\link{kernelFun}} for bare-bones kernels.
}
