% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing-functions.R
\name{semiParLM}
\alias{semiParLM}
\title{Robinson’s semi-parametric regression}
\usage{
semiParLM(
  xnames,
  znames,
  yname,
  data,
  bw.x = NULL,
  bw.y = NULL,
  order = 4,
  ...
)
}
\arguments{
\item{xnames}{A character vector of exogenous regressors entering the linear part of the model.}

\item{znames}{A character vector of exogenous regressors entering the non-parametric part of the model. Must be continuous variables (an assumption made by Robinson)}

\item{yname}{A character scalar: dependent variable name.}

\item{data}{A data frame containing the variables specified in \code{xnames}, \code{znames}, \code{yname}.}

\item{bw.x}{A list containing bandwidth vectors or scalars for smoothing X1(Z), X2(Z), ...}

\item{bw.y}{A numeric vector or scalar for smoothing Y(Z).}

\item{order}{Non-parametric smoothing order. Robinson (1988) inststs that \code{order >= length(znames)}.}

\item{...}{Passed to \code{kernelSmooth}.}
}
\value{
Returns an object of class "lm", like an ordinary \code{lm()}, would, with an extra component: \code{non.parametric} for the estimated mu(Z).
}
\description{
Robinson’s semi-parametric regression
}
\examples{
set.seed(1)
n <- 1000
age <- rchisq(n, 5)
female <- rbinom(n, 1, 0.5)
muscle <- rnorm(n, sd = 3) - female
knowledge <- runif(n, 1, 8) + age/2 + female - muscle/4
mu <- function(x1, x2) 0.5*x1 + 0.5*x2 + 0.02*x1*x2 + 2*sin(x1) + 2*cos(0.5*x2)
wage <- 1 + age + female + age*female + mu(muscle, knowledge) + rnorm(n, sd = 2)
d <- data.frame(cbind(wage, age, female, age.female = age*female, muscle, knowledge))
diag(var(d))
plot(d)
x1grid <- quantile(muscle, 1:49/50)
x2grid <- quantile(knowledge, 1:49/50)
persp(x1grid, x2grid, outer(x1grid, x2grid, mu), theta = 45, phi = 30, ticktype = "detailed")

mod.linear <- lm(wage ~ age + female + age*female + muscle + I(muscle^2) + knowledge +
  I(knowledge^2), data = d)
mod.semipar <- semiParLM(xnames = c("age", "female", "age.female"),
  znames = c("muscle", "knowledge"), yname = "wage", data = d)

plot(d$age - mod.semipar$predicted.X[, 1], d$wage - mod.semipar$predicted.Y,
  col = c("blue", "red")[d$female + 1])
plot(scale(d$muscle, scale = FALSE), scale(mu(muscle, knowledge), scale = FALSE))
points(scale(d$muscle, scale = FALSE), scale(mod.semipar$non.parametric, scale = FALSE),
  col = "red", pch = 16, cex = 0.5)
plot(scale(d$knowledge, scale = FALSE),   scale(mu(muscle, knowledge),      scale = FALSE))
points(scale(d$knowledge, scale = FALSE), scale(mod.semipar$non.parametric, scale = FALSE),
  col = "red", pch = 16, cex = 0.5)
}
